Moved RcppArmadilloConfig.h
