+++ inst/include/RcppArmadillo/config/RcppArmadilloConfig.h
